deid_assess_platform/requirements.txt
PySide6==6.7.2
polars==1.9.0
pandas==2.2.2
python-docx==1.1.2
matplotlib==3.9.2
plotly==5.24.1
